board-service 수정 및 삭제 구현 auth-service 로그인시 admin일때 회원관리 표시

+++ board-service/src/main/java/org/kosa/boardservice/dto/BoardDto.java
@@ -16,7 +16,7 @@ import java.time.format.DateTimeFormatter;
 @NoArgsConstructor
 public class BoardDto {
 
-    private Integer bno;
+    private Long bno;
     private String title;
     private String content;
     private String writer;
